Fetch a "secret"
For the moment, this is all hard-coded to talk to the localstack version of SecretsManager. But one step at a time.

In order to get to the SecretsManager, this pulls in the AWS SDK. It needs a newer version than Querki has, which conflicts with Play's version of Netty. So we turn off Netty and turn on akka-http -- a little premature in Play 2.5, but hopefully functional enough.