Proof of concept is now fully working
Main change was to give up on Artery for now, and back off to using Netty for the clustering. To my surprise and relief, this does not seem to clash with AWS or Play.

Also needed to introduce DData for now, so that Cluster Sharding can use it. (Not needed for Querki, since that uses Persistence.)

Some small enhancements to the business logic, to prove that everything is clustering correctly.

Woohoo!